# openapi.json
{
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "definitions": {
...
    "ErrorResponse": {
      "properties": {
        "error": {
          "type": "string"
        },
        "success": {
          "type": "boolean"
        }
      },
      "type": "object"
    },
...
    "TrafficStop": {
      "properties": {
        "lat": {
          "type": "number"
        },
        "lng": {
          "type": "number"
        }
      },
      "required": [
        "lat",
        "lng"
      ],
      "type": "object"
    },
...
  },
  "info": {
    "contact": {
      "name": "RouteForce Support",
      "url": "https://github.com/ApacheEcho/RouteForceRouting"
    },
    "description": "API documentation for RouteForce Routing application - A comprehensive route optimization platform for field execution teams",
    "title": "RouteForce Routing API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/traffic/optimize": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "constraints": {
                  "type": "object"
                },
                "start_location": {
                  "$ref": "#/definitions/TrafficStop"
                },
                "stores": {
                  "items": {
                    "$ref": "#/definitions/TrafficStop"
                  },
                  "type": "array"
                }
              },
              "required": [
                "stores"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Traffic-optimized route returned",
            "examples": {
              "application/json": {
                "algorithm_used": "traffic_optimized",
                "metrics": {
                  "processing_time": 0.12,
                  "total_distance": 12.3,
                  "total_duration": 1800,
                  "traffic_delay": 180
                },
                "route": [
                  {
                    "lat": 40.71,
                    "lng": -74.0,
                    "name": "Store A"
                  },
                  {
                    "lat": 40.76,
                    "lng": -73.98,
                    "name": "Store B"
                  }
                ],
                "success": true,
                "traffic_data": {
                  "delay_minutes": 3
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "summary": "Generate traffic-optimized route",
        "tags": [
          "Traffic"
        ]
      }
    },
    "/api/traffic/predict": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "future_hours": {
                  "default": [
                    1,
                    2,
                    4,
                    8
                  ],
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                "stores": {
                  "items": {
                    "$ref": "#/definitions/TrafficStop"
                  },
                  "type": "array"
                }
              },
              "required": [
                "stores"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Traffic prediction available",
            "examples": {
              "application/json": {
                "best_time": {
                  "hour": 1
                },
                "predictions": [
                  {
                    "delay_minutes": 2,
                    "hour": 1
                  },
                  {
                    "delay_minutes": 5,
                    "hour": 2
                  }
                ],
                "processing_time": 0.06,
                "recommendation": {
                  "depart_after_hours": 1
                },
                "success": true
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "summary": "Predict traffic conditions for future departure times",
        "tags": [
          "Traffic"
        ]
      }
    },
    "/api/traffic/status": {
      "get": {
        "responses": {
          "200": {
            "description": "Service status returned",
            "examples": {
              "application/json": {
                "api_key_configured": true,
                "available": true,
                "cache_stats": {
                  "hits": 12,
                  "misses": 3
                },
                "config": {
                  "avoid_ferries": false,
                  "avoid_highways": false,
                  "avoid_tolls": false,
                  "cache_duration": 3600,
                  "max_waypoints": 23,
                  "traffic_model": "best_guess"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "summary": "Get status of traffic service and Google Maps API",
        "tags": [
          "Traffic"
        ]
      }
    },
...
  },
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "in": "header",
      "name": "X-API-Key",
      "type": "apiKey"
    },
    "JWT": {
      "description": "JWT Bearer token. Use: 'Bearer <token>'",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0",
...
}